requirements.txt
pandas==2.2.2
gspread==6.0.2
google-auth==2.29.0
requests==2.32.3
